Initial modifications uploaded
Made a bunch of changes to make PlainTask look and work a bit more like
a text version of OmniFocus. Only Tested on a Mac:

Usability
- Return creates a new task. ⌘+Return will do a carriage return (or
else use Enter on the numpad).
- Changed scope in .tmlanguage for tags. Now should be able to
highlight multiple tags on the same line.
- Added scope for @due tasks

Cosmetic:
- Added breathing room above and below task lines.
- Changed font to Lucida Grande (which will not allow for bold or
italics, only works on monospaced fonts)
- Changed completed task marked to box with check.

Haven't updated readme at all. This is the first pass of changes.